SystemUI: Introduce user interface for Alert Sliders
Ported from OxygenOS and reworked for our alert slider

implementation. We target AudioManager instead of Zen,

icons are also the same as aosp and the dialog uses the

material theme as well as support for our themes.

To use, the alert slider config must be enabled. By default,

the dialog shows on the left side. To move it to the right

side, set the location config to 1.

Change-Id: Ie1954a44cc5242c95a731abd7404379ea638fe70